backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
requests
tweepy
beautifulsoup4
langchain
langchain-openai
websockets
cryptography
pydantic
pydantic-settings
